Added abstract baseSplitActivity class to better handle attachBaseContext.
Moved attachBaseContext for the Activity in dyn-modules in a base abstract
Activity class that needs to be used as base.  This avoid to have to repeat the
same code for the first activity launched in a dynamic-module.

Updated README with a reference to the BaseSplitActivity.

The project has been updated to work with Android Studio v3.2.

Change-Id: I1bf6c88b4e8707fc0638b8f5e9889d069112c0cc